型の誤りを訂正
	modified:   hypervisor/src/thread/thread_local.h
	modified:   hypervisor/src/thread/thread_setup_array.c

+++ hypervisor/src/thread/thread_local.h
@@ -33,7 +33,7 @@ extern "C" {
 
 /* functions */
 
-uint64_t *thread_setup_array(uint32_t no);
+uintptr_t *thread_setup_array(uint32_t no);
 
 #ifdef __cplusplus
 }
